+++ components/zeekdatabase/src/virtualtablemodule.cpp
@@ -277,14 +277,15 @@ int VirtualTableModule::onTableColumn(sqlite3_vtab_cursor *cursor,
     const auto &current_column_data =
         std::get<std::int64_t>(current_column_value_data);
 
-    sqlite3_result_int(context, current_column_data);
+    sqlite3_result_int(context, static_cast<int>(current_column_data));
 
   } else if (std::holds_alternative<std::string>(current_column_value_data)) {
     const auto &current_column_data =
         std::get<std::string>(current_column_value_data);
 
     sqlite3_result_text(context, current_column_data.c_str(),
-                        current_column_data.size(), SQLITE_STATIC);
+                        static_cast<int>(current_column_data.size()),
+                        SQLITE_STATIC);
 
   } else if (std::holds_alternative<double>(current_column_value_data)) {
     const auto &current_column_data =
